+++ Assets/Scripts/Interfaces/IBreakable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IBreakable
 {
-    public void Damage(GameObject sender, int damage, BreakableType type, int toughness);
+    public void Damage(GameObject sender, float damage, BreakableType type, int toughness);
 
 }
